requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.3
multidict==4.7.6
pycparser==2.21
PyNaCl==1.5.0
yarl==1.8.1
